third_party/blink/web_tests/fast/dom/Window/property-access-in-closure-after-navigation-expected.txt

Test that an inner global captured in a closure acts as if its creating context is detached when the creating context is navigated.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window is self
PASS window is frames
PASS parent is top
PASS window is non-null.
PASS self is non-null.
PASS frames is non-null.
PASS parent is non-null.
PASS top is non-null.
